+++ packages/device_info_plus/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## NEXT
 
 * Add 2 integration test cases.
+* Update integration tests.
 
 ## 1.4.0