+++ packages/flutter_tts/CHANGELOG.md
@@ -1,3 +1,7 @@
+## NEXT
+
+* Update integration tests.
+
 ## 1.7.0
 
 * Return a boolean from `isLanguageAvailable` to match the behavior of other platforms.