+++ packages/flutter_webrtc/CHANGELOG.md
@@ -1,3 +1,7 @@
+## NEXT
+
+* Update integration tests.
+
 ## 0.2.3
 
 * Fix `RTCPeerConnection.close()` to be idempotent; calling it more than once no longer throws an error.